Got annotator of extract working.
