+++ frontend/src/graphql/cache.ts
@@ -219,6 +219,9 @@ export const filterToAnnotationLabelId = makeVar<string>(""); // Not used elsewh
 export const selectedAnnotation = makeVar<ServerAnnotationType | null>(null);
 export const displayAnnotationOnAnnotatorLoad =
   makeVar<ServerAnnotationType | null>(null);
+export const onlyDisplayTheseAnnotations = makeVar<
+  ServerAnnotationType[] | undefined
+>(undefined);
 export const annotationContentSearchTerm = makeVar<string>("");
 export const selectedMetaAnnotationId = makeVar<string>("");
 export const includeStructuralAnnotations = makeVar<boolean>(false); // These are weird as they don't have a labelset and user probably doesn't want to see them.
