+++ opencontractserver/tasks/extract_tasks.py
@@ -112,7 +112,10 @@ def llama_index_doc_query(cell_id, similarity_top_k=3):
         retriever = index.as_retriever(similarity_top_k=similarity_top_k)
 
         results = retriever.retrieve(search_text if search_text else query)
-        retrieved_annotation_ids = [n.node.extra_info['id'] for n in results]
+        for r in results:
+            print(f"Result {r.node.extra_info['label_id']}:\n{r}")
+        retrieved_annotation_ids = [n.node.extra_info['label_id'] for n in results]
+        print(f"retrieved_annotation_ids: {retrieved_annotation_ids}")
         datacell.sources.add(*retrieved_annotation_ids)
 
         retrieved_text = "\n".join(
